+++ CBOR/PeterO/Cbor/CBORDataUtilitiesTextString.cs
@@ -13,8 +13,8 @@ namespace PeterO.Cbor
 {
     internal static class CBORDataUtilitiesTextString
     {
-        private const long DoubleNegInfinity = unchecked((long)(0xfffL << 52));
-        private const long DoublePosInfinity = unchecked((long)(0x7ffL << 52));
+        private const long DoubleNegInfinity = unchecked(0xfffL << 52);
+        private const long DoublePosInfinity = unchecked(0x7ffL << 52);
 
         internal static CBORObject ParseJSONNumber(
           string chars,
@@ -269,7 +269,7 @@ namespace PeterO.Cbor
                 int vi = numOffset;
                 for (; vi < endPos; ++vi)
                 {
-                    v = (v * 10) + (int)(chars[vi] - '0');
+                    v = (v * 10) + (chars[vi] - '0');
                 }
                 if ((v != 0 || !negative) && v < (1L << 53) - 1)
                 {
@@ -323,7 +323,7 @@ namespace PeterO.Cbor
                         {
                             ++digitCount;
                         }
-                        lv = checked((lv * 10) + (int)(chars[vi] - '0'));
+                        lv = checked((lv * 10) + (chars[vi] - '0'));
                     }
                     for (vi = decimalPointPos + 1; vi < endPos; ++vi)
                     {
@@ -336,7 +336,7 @@ namespace PeterO.Cbor
                         {
                             ++digitCount;
                         }
-                        lv = checked((lv * 10) + (int)(chars[vi] - '0'));
+                        lv = checked((lv * 10) + (chars[vi] - '0'));
                     }
                     if (negative)
                     {
