+++ CBOR/PeterO/Cbor/CBORNumberExtra.cs
@@ -161,7 +161,7 @@ namespace PeterO.Cbor
             }
             if (this.IsZero())
             {
-                return (ushort)0;
+                return 0;
             }
             if (this.IsNegative())
             {
