+++ CBORTest/StringOutput.cs
@@ -36,7 +36,7 @@ namespace Test
             {
                 if (str.Length == 1)
                 {
-                    this.WriteCodePoint((int)str[0]);
+                    this.WriteCodePoint(str[0]);
                 }
                 else
                 {
@@ -67,7 +67,7 @@ namespace Test
             {
                 if (length == 1)
                 {
-                    this.WriteCodePoint((int)str[index]);
+                    this.WriteCodePoint(str[index]);
                 }
                 else
                 {
@@ -131,7 +131,7 @@ namespace Test
                 for (var i = 0; i < length; ++i)
                 {
                     byte b = bytes[i + index];
-                    if ((((int)b) & 0x7f) != b)
+                    if ((b & 0x7f) != b)
                     {
                         throw new ArgumentException("str is non-ASCII");
                     }
